#27 - グループ一括削除、グループ名変更機能が動くようにした。 - アカウントリスト画面をコンポーネント分けした（リファクタリング）
