+++ .eslintrc.json
@@ -2,7 +2,11 @@
   "env": {
     "browser": true
   },
-  "extends": ["airbnb", "plugin:vue/recommended", "prettier"],
+  "extends": [
+    "airbnb",
+    "plugin:vue/recommended",
+    "prettier"
+  ],
   "globals": {
     "Vue": true,
     "chrome": true,
@@ -13,11 +17,15 @@
     "ecmaVersion": 11,
     "sourceType": "module"
   },
-  "plugins": ["vue"],
+  "plugins": [
+    "vue"
+  ],
   "rules": {
     "linebreak-style": "off",
     "no-console": "off",
-    "no-underscore-dangle":"off",
-    "no-param-reassign": "off"
+    "no-underscore-dangle": "off",
+    "no-param-reassign": "off",
+    "no-else-return": "off",
+    "html-self-closing": "off"
   }
 }
